drm/msm/sde: Skip unneeded register reads when getting line count
More often than not, get_vsync_info() is used to only get the write line
count, while the other values it returns are left unused. This is not
optimal since it is done on every display commit. We can eliminate the
superfluous register reads by adding a parameter specifying if only the
R/W line count is requested.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
Change-Id: I3cb8a6e8e5454e5cf1f726b8e42946ce0c808746
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>